move all notification tests to consistent package
