+++ android/src/test/java/com/thebluealliance/androidclient/gcm/notifications/AllianceSelectionNotificationTest.java
@@ -1,4 +1,4 @@
-package com.thebluealliance.androidclient.notifications;
+package com.thebluealliance.androidclient.gcm.notifications;
 
 import android.app.Notification;
 import android.content.Context;
@@ -9,59 +9,68 @@ import com.google.gson.JsonParseException;
 import com.thebluealliance.androidclient.R;
 import com.thebluealliance.androidclient.activities.ViewEventActivity;
 import com.thebluealliance.androidclient.adapters.ViewEventFragmentPagerAdapter;
+import com.thebluealliance.androidclient.database.writers.EventWriter;
 import com.thebluealliance.androidclient.datafeed.framework.ModelMaker;
+import com.thebluealliance.androidclient.gcm.notifications.AllianceSelectionNotification;
 import com.thebluealliance.androidclient.gcm.notifications.NotificationTypes;
-import com.thebluealliance.androidclient.gcm.notifications.ScheduleUpdatedNotification;
 import com.thebluealliance.androidclient.helpers.MyTBAHelper;
+import com.thebluealliance.androidclient.models.Event;
 import com.thebluealliance.androidclient.models.StoredNotification;
 
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mockito.Mock;
 import org.robolectric.RuntimeEnvironment;
-
-import java.text.DateFormat;
-import java.util.Date;
 
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
+import static org.mockito.Matchers.anyLong;
+import static org.mockito.Matchers.eq;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
 
 @RunWith(AndroidJUnit4.class)
-public class ScheduleUpdatedNotificationTest {
-    private Context mContext;
-    private ScheduleUpdatedNotification mNotification;
+public class AllianceSelectionNotificationTest {
+
+    @Mock private Context mContext;
+    @Mock private EventWriter mWriter;
+    private AllianceSelectionNotification mNotification;
     private JsonObject mData;
 
     @Before
     public void setUp() {
         mContext = RuntimeEnvironment.application.getApplicationContext();
-        mData = ModelMaker.getModel(JsonObject.class, "notification_schedule_updated");
-        mNotification = new ScheduleUpdatedNotification(mData.toString());
+        mWriter = mock(EventWriter.class);
+        mData = ModelMaker.getModel(JsonObject.class, "notification_alliance_selection");
+        mNotification = new AllianceSelectionNotification(mData.toString(), mWriter);
     }
 
     @Test
     public void testParseData() {
         mNotification.parseMessageData();
 
-        assertEquals(mNotification.getEventKey(), "2015ausy");
-        assertEquals(mNotification.getEventName(), "Australia Regional");
-        assertNotNull(mNotification.getMatchTime());
-        assertEquals(mNotification.getMatchTime().getAsInt(), 1397330280);
+        assertEquals(mNotification.getEventKey(), "2014necmp");
+
+        Event event = mNotification.getEvent();
+        assertNotNull(event);
     }
 
-    @Test(expected = JsonParseException.class)
-    public void testNoEventKey() {
-        mData.remove("event_key");
-        mNotification = new ScheduleUpdatedNotification(mData.toString());
+    @Test
+    public void testDbWrite() {
         mNotification.parseMessageData();
+        mNotification.updateDataLocally();
+
+        Event event = mNotification.getEvent();
+        verify(mWriter).write(eq(event), anyLong());
     }
 
     @Test(expected = JsonParseException.class)
-    public void testNoEventName() {
-        mData.remove("event_name");
-        mNotification = new ScheduleUpdatedNotification(mData.toString());
+    public void testNoEvent() {
+        mData.remove("event");
+        mNotification = new AllianceSelectionNotification(mData.toString(), mWriter);
         mNotification.parseMessageData();
     }
 
@@ -71,16 +80,11 @@ public class ScheduleUpdatedNotificationTest {
         Notification notification = mNotification.buildNotification(mContext, null);
         assertNotNull(notification);
 
-        long scheduledStartTimeUNIX = mNotification.getMatchTime().getAsLong();
-        Date scheduledStartTime = new Date(scheduledStartTimeUNIX * 1000);
-        DateFormat format = android.text.format.DateFormat.getTimeFormat(mContext);
-        String startTime = format.format(scheduledStartTime);
-
         StoredNotification stored = mNotification.getStoredNotification();
         assertNotNull(stored);
-        assertEquals(stored.getType(), NotificationTypes.SCHEDULE_UPDATED);
-        assertEquals(stored.getTitle(), mContext.getString(R.string.notification_schedule_updated_title, "AUSY"));
-        assertEquals(stored.getBody(), mContext.getString(R.string.notification_schedule_updated_with_time, "Australia", startTime));
+        assertEquals(stored.getType(), NotificationTypes.ALLIANCE_SELECTION);
+        assertEquals(stored.getTitle(), mContext.getString(R.string.notification_alliances_updated_title, "NECMP"));
+        assertEquals(stored.getBody(), mContext.getString(R.string.notification_alliances_updated, "New England"));
         assertEquals(stored.getMessageData(), mData.toString());
         assertEquals(stored.getIntent(), MyTBAHelper.serializeIntent(mNotification.getIntent(mContext)));
         assertNotNull(stored.getTime());
@@ -93,6 +97,6 @@ public class ScheduleUpdatedNotificationTest {
         assertNotNull(intent);
         assertEquals(intent.getComponent().getClassName(), "com.thebluealliance.androidclient.activities.ViewEventActivity");
         assertEquals(intent.getStringExtra(ViewEventActivity.EVENTKEY), mNotification.getEventKey());
-        assertEquals(intent.getIntExtra(ViewEventActivity.TAB, -1), ViewEventFragmentPagerAdapter.TAB_MATCHES);
+        assertEquals(intent.getIntExtra(ViewEventActivity.TAB, -1), ViewEventFragmentPagerAdapter.TAB_ALLIANCES);
     }
 }
